backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
Django==5.2
django-constance==4.3.2
django-cors-headers==4.7.0
django-multiselectfield==0.1.13
djangorestframework==3.16.0
filelock==3.18.0
identify==2.6.9
idna==3.10
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pre_commit==4.2.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
python-decouple==3.8
PyYAML==6.0.2
qrcode==8.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.1
urllib3==2.3.0
virtualenv==20.30.0
